Hardened security of webauthn flow

+++ src/main/java/constants/Const.java
@@ -11,5 +11,5 @@ public final class Const {
     public static final String COOKIE_NAME = "karakal-auth";
     public static final String KARAKAL_CACHE_NAME = "karakal-auth-cache";
     public static final String DASHBOARD = "dashboard";
-    public static final long COOKIE_MAX_AGE = 600L;
+    public static final int COOKIE_MAX_AGE = 600;
 }
